script parsing should be more loose on pushdata
This testnet transaction was being parsed incorrectly:

cc64de74ba7002bbf4e3646824d7bbf0920004fb2ce45aa7270c4116ff11b715

Script was throwing an error when it should not have been. The error was that
PUSHDATA1 was trying to push 117 bytes to the stack, but it was followed by
only 75 bytes. But this transaction is accepted as valid by bitcoin-qt on
testnet. So we are mistaken by throwing an error in this case.