remove network stuff from main build
The network classes Connection, Peer and PeerManager are not useful from the
browser, so I have removed them from the main browser bundle by default.  This
saves several tens of kilobytes from the browser bundle (since not only are
these files not included, but their dependencies are not included either).