Add feePerByte for fee calculation options
Bitcoin cash allows for 1 sat/byte fees.
The current feePerKb forces the user that is paying, to pay too much
most of the time.
This keeps the default as feePerKb since devs might assume this is the
default behaviour as is currently the case. This assumption can be
observed in tests as well.